- 관리자 > 대시보드 > 승선현황 api

+++ src/main/java/com/tobe/fishking/v2/model/admin/DashBoardManageDto.java
@@ -60,10 +60,10 @@ public class DashBoardManageDto {
     @AllArgsConstructor
     @Setter
     public class RideData{
-        public Integer waitRideCount;
-        public Integer confirmRide;
-        public Integer failConfirm;
-        public Integer cancelRide;
+        public Long waitRideCount;
+        public Long confirmRide;
+        public Long failConfirm;
+        public Long cancelRide;
     }
 
     private List<NoticeData> noticeData;
